my_pytest_playwright_project/requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
aspose-diagram==25.3.0
aspose-diagram-python==25.3
attrs==24.2.0
Authlib==1.5.2
azure-core==1.32.0
azure-devops==7.1.0b4
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==6.2.0
certifi==2025.1.31
cffi==1.17.0
charset-normalizer==3.4.1
choreographer==1.1.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.9
contourpy==1.3.3
cryptography==44.0.0
cycler==0.12.1
deepdiff==8.4.2
defusedxml==0.7.1
dparse==0.6.4
et_xmlfile==2.0.0
Faker==23.3.0
filelock==3.16.1
Flask==3.1.1
fonttools==4.59.1
fpdf2==2.8.4
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.2
h11==0.16.0
hl7apy==1.3.4
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
jpype1==1.5.2
jsonschema==4.23.0
jsonschema-path==0.3.3
jsonschema-specifications==2024.10.1
kaleido==1.1.0
kiwisolver==1.4.9
lazy-object-proxy==1.10.0
logistro==1.1.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
matplotlib==3.10.5
mdurl==0.1.2
MouseInfo==0.1.3
msrest==0.7.1
narwhals==2.5.0
natsort==8.4.0
nltk==3.9.1
numpy==2.2.5
oauthlib==3.2.2
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
opencv-python==4.12.0.88
openpyxl==3.1.5
orderly-set==5.4.1
orjson==3.11.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.3.2
paramiko==3.5.0
pathable==0.4.4
pbr==6.1.1
pefile==2023.2.7
pillow==11.1.0
platformdirs==4.3.8
playwright==1.52.0
plotly==6.3.0
pluggy==1.6.0
plum-dispatch==1.7.4
protobuf==6.32.1
psutil==6.1.1
pyarrow==21.0.0
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pyee==13.0.0
PyGetWindow==0.0.9
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
PyMsgBox==2.0.1
PyNaCl==1.5.0
pyodbc==5.2.0
pyparsing==3.2.3
pyperclip==1.10.0
PyRect==0.2.0
PyScreeze==1.0.1
pysftp==0.2.9
PySocks==1.7.1
pytest==8.4.2
pytest-timeout==2.4.0
python-dateutil==2.8.2
python-docx==1.2.0
python-dotenv==1.1.0
pytoolconfig==1.3.1
pytweening==1.2.0
pytz==2025.2
pywin32==308
pywin32-ctypes==0.2.3
pywinauto==0.6.9
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
reportlab==4.4.3
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.0.0
rope==1.13.0
rpds-py==0.22.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safety-schemas==0.0.14
seaborn==0.13.2
selenium==4.28.1
setuptools==75.8.0
shellingham==1.5.4
simplejson==3.20.2
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spire==0.4.2
Spire.Xls==14.12.0
stevedore==5.4.1
streamlit==1.50.0
svgwrite==1.4.3
tenacity==9.1.2
tk==0.1.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.5.2
tqdm==4.67.1
trio==0.26.2
trio-websocket==0.11.1
typer==0.15.4
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
watchdog==6.0.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
